add determinePoolDisplay to utils
